requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.8.0
ffmpeg-python>=0.2.0
faster-whisper>=1.0.3
pyav>=12.1.0
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.20.0
tqdm>=4.65.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib